Spawn a shell to run commands in instead for running commands in a spawned shell.
